rename file to match component
