+++ components/ProposalStateBadge.tsx
