+++ pages/dao/[symbol]/proposal/[pk]/index.tsx
@@ -2,7 +2,7 @@ import ReactMarkdown from 'react-markdown/react-markdown.min'
 import remarkGfm from 'remark-gfm'
 import { ExternalLinkIcon } from '@heroicons/react/outline'
 import useProposal from 'hooks/useProposal'
-import ProposalStateBadge from 'components/ProposalStatusBadge'
+import ProposalStateBadge from '@components/ProposalStateBadge'
 import { InstructionPanel } from 'components/instructions/instructionPanel'
 import DiscussionPanel from 'components/chat/DiscussionPanel'
 import VotePanel from 'components/VotePanel'
